# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "KB API",
        "description": "HTTP JSON API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api"
        }
    ],
    "paths": {
        "/auth/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Получение токена по логину и паролю",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/LoginRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Токен аутентификации",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiTokenResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Неаутентифицирован"
                    }
                },
                "security": []
            }
        },
...
        "/auth/azure/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Получение токена после перенаправления",
                "description": "Нужно пробросить в тело все параметры с query string редиректа 'как есть'",
                "responses": {
                    "200": {
                        "description": "Токен аутентификации",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiTokenResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Неаутентифицирован"
                    }
                },
                "security": []
            }
        },
        "/auth/me": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "Получение данных текущего пользователя",
                "responses": {
                    "200": {
                        "description": "Данные текущего пользователя",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Неаутентифицирован"
                    }
                },
                "security": {
                    "apiKey": {}
                }
            }
        },
...
        "/sections": {
            "get": {
                "tags": [
                    "Sections"
                ],
                "summary": "Получение списка разделов",
                "responses": {
                    "200": {
                        "description": "Список разделов",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#components/schemas/SectionResource"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Неаутентифицирован"
                    }
                },
                "security": {
                    "apiKey": {}
                }
            },
            "post": {
                "tags": [
                    "Sections"
                ],
                "summary": "Создание раздела (admin, moderator)",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SectionCreateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Созданный раздел",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#components/schemas/SectionWithFieldsResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Неаутентифицирован"
                    }
                },
                "security": {
                    "apiKey": {}
                }
            }
        },
        "/sections/{section}": {
            "get": {
                "tags": [
                    "Sections"
                ],
                "summary": "Получение раздела с его полями",
                "parameters": [
                    {
                        "name": "section",
                        "in": "path",
                        "description": "Идентификатор раздела"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Раздел с его значениями",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#components/schemas/SectionWithFieldsResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Неаутентифицирован"
                    }
                },
                "security": {
                    "apiKey": {}
                }
            },
            "delete": {
                "tags": [
                    "Sections"
                ],
                "summary": "Удаление раздела (admin, moderator)",
                "parameters": [
                    {
                        "name": "section",
                        "in": "path",
                        "description": "Идентификатор раздела"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Удалено"
                    },
                    "401": {
                        "description": "Неаутентифицирован"
                    }
                },
                "security": {
                    "apiKey": {}
                }
            }
        },
        "/sections/{sections}": {
            "patch": {
                "tags": [
                    "Sections"
                ],
                "summary": "Обновление раздела (admin, moderator)",
                "parameters": [
                    {
                        "name": "section",
                        "in": "path",
                        "description": "Идентификатор раздела"
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SectionUpdateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Обновленный раздел",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#components/schemas/SectionWithFieldsResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Неаутентифицирован"
                    }
                },
                "security": {
                    "apiKey": {}
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "LoginRequest": {
                "properties": {
                    "login": {
                        "type": "string",
                        "example": "user@some.com"
                    },
                    "password": {
                        "type": "string",
                        "example": "123123"
                    }
                },
                "type": "object"
            },
...
            "SectionCreateRequest": {
                "required": [
                    "title",
                    "is_dictionary",
                    "is_navigation"
                ],
                "properties": {
                    "id": {
                        "type": "string",
                        "example": "123e4567-e89b-12d3-a456-426655440000"
                    },
                    "title": {
                        "type": "string",
                        "maximum": "255",
                        "minimum": "1",
                        "example": "Имя раздела"
                    },
                    "is_dictionary": {
                        "type": "boolean"
                    },
                    "is_navigation": {
                        "type": "boolean"
                    },
                    "image": {
                        "nullable": true,
                        "type": "string",
                        "maximum": "255",
                        "minimum": "1",
                        "example": "Имя раздела"
                    },
                    "fields": {
                        "type": "array",
                        "items": {
                            "required": [
                                "title",
                                "type",
                                "required",
                                "is_present_in_card",
                                "is_filterable"
                            ],
                            "properties": {
                                "id": {
                                    "type": "string",
                                    "example": "123e4567-e89b-12d3-a456-426655440000"
                                },
                                "title": {
                                    "type": "string"
                                },
                                "description": {
                                    "type": "string"
                                },
                                "sort_index": {
                                    "type": "integer"
                                },
                                "required": {
                                    "type": "boolean"
                                },
                                "is_present_in_card": {
                                    "type": "boolean"
                                },
                                "is_filterable": {
                                    "type": "boolean"
                                },
                                "type": {
                                    "oneOf": [
                                        {
                                            "$ref": "#components/schemas/TypeString"
                                        },
                                        {
                                            "$ref": "#components/schemas/TypeInteger"
                                        },
                                        {
                                            "$ref": "#components/schemas/TypeText"
                                        },
                                        {
                                            "$ref": "#components/schemas/TypeWiki"
                                        },
                                        {
                                            "$ref": "#components/schemas/TypeFloat"
                                        },
                                        {
                                            "$ref": "#components/schemas/TypeBoolean"
                                        },
                                        {
                                            "$ref": "#components/schemas/TypeEnum"
                                        },
                                        {
                                            "$ref": "#components/schemas/TypeFile"
                                        },
                                        {
                                            "$ref": "#components/schemas/TypeList"
                                        }
                                    ]
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            },
            "SectionUpdateRequest": {
                "properties": {
                    "id": {
                        "type": "string",
                        "example": "123e4567-e89b-12d3-a456-426655440000"
                    },
                    "title": {
                        "type": "string",
                        "maximum": "255",
                        "minimum": "1",
                        "example": "Имя раздела"
                    },
                    "is_dictionary": {
                        "type": "boolean"
                    },
                    "is_navigation": {
                        "type": "boolean"
                    },
                    "image": {
                        "nullable": true,
                        "type": "string",
                        "maximum": "255",
                        "minimum": "1",
                        "example": "Имя раздела"
                    },
                    "fields": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "id": {
                                    "type": "string",
                                    "example": "123e4567-e89b-12d3-a456-426655440000"
                                },
                                "title": {
                                    "type": "string"
                                },
                                "description": {
                                    "type": "string"
                                },
                                "sort_index": {
                                    "type": "integer"
                                },
                                "required": {
                                    "type": "boolean"
                                },
                                "is_present_in_card": {
                                    "type": "boolean"
                                },
                                "type": {
                                    "oneOf": [
                                        {
                                            "$ref": "#components/schemas/TypeString"
                                        },
                                        {
                                            "$ref": "#components/schemas/TypeInteger"
                                        },
                                        {
                                            "$ref": "#components/schemas/TypeText"
                                        },
                                        {
                                            "$ref": "#components/schemas/TypeWiki"
                                        },
                                        {
                                            "$ref": "#components/schemas/TypeFloat"
                                        },
                                        {
                                            "$ref": "#components/schemas/TypeBoolean"
                                        },
                                        {
                                            "$ref": "#components/schemas/TypeEnum"
                                        },
                                        {
                                            "$ref": "#components/schemas/TypeFile"
                                        },
                                        {
                                            "$ref": "#components/schemas/TypeList"
                                        }
                                    ]
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            },
...
            "ApiTokenResource": {
                "properties": {
                    "token": {
                        "type": "string",
                        "example": "access_token"
                    }
                },
                "type": "object"
            },
...
            "UserResource": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": "123"
                    },
                    "name": {
                        "nullable": true,
                        "type": "string",
                        "example": "Ульянов Владимир Ильич"
                    },
                    "avatar": {
                        "nullable": true,
                        "type": "string",
                        "example": "absolute/path/to/image.jpg"
                    },
                    "role": {
                        "type": "string",
                        "enum": [
                            "admin",
                            "moderator",
                            "user"
                        ]
                    },
                    "login": {
                        "type": "string",
                        "example": "john15"
                    },
                    "email": {
                        "nullable": true,
                        "type": "string",
                        "example": "user@mail.com"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "apiKey": {
                "type": "apiKey",
                "description": "Api key",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "security": [
        {
            "apiKey": []
        }
    ]
}